Revert "Move and rename jobs"
This reverts commit f489377fb63beb9f5edee56c4c791a3d9be492ec.
