--- app/workers/streaming_history/save_track_job.rb
@@ -1,22 +1,0 @@
-require 'rspotify'
-
-module StreamingHistory
-  class SaveTrackJob
-    include Sidekiq::Worker
-
-    def perform(json_listen, user_id)
-      listen = JSON.parse(json_listen)
-
-      track = RSpotify::Track.search(listen['trackName'] + ' ' + listen['artistName']).first
-      Scrobble.create(
-          track_id: track.id,
-          created_at: listen['endTime'],
-          user_id: user_id,
-          updated_at: listen['endTime'],
-          artist_ids: track.artists.map(&:id),
-          track_name: listen['trackName'],
-          artist_name: listen['artistName']
-      ) if track
-    end
-  end
-end
